+++ test/test_migration_planner.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 require 'minitest/autorun'
+require 'tmpdir'
+require 'fileutils'
 require_relative '../lib/scene_texture_switcher/legacy_library_scanner'
 require_relative '../lib/scene_texture_switcher/migration_planner'
 
@@ -22,5 +24,20 @@ class MigrationPlannerTest < Minitest::Test
     assert_equal false, plan[:writable]
     assert_empty plan[:conflicts]
   end
-end
 
+  def test_numeric_working_files_are_preserved_without_becoming_textures
+    Dir.mktmpdir do |root|
+      folder = File.join(root, 'Surface01')
+      FileUtils.mkdir_p(folder)
+      File.write(File.join(folder, '01.png'), 'texture')
+      File.write(File.join(folder, '01.psd'), 'working file')
+      scan = SceneTextureSwitcher::LegacyLibraryScanner.new(root).scan
+
+      plan = SceneTextureSwitcher::MigrationPlanner.new(scan).plan
+
+      assert_includes plan[:mappings].map { |mapping| mapping[:destination_relative] }, File.join('01', 'Surface01.psd')
+      assert_equal 1, plan[:supporting_file_count]
+      assert_empty plan[:conflicts]
+    end
+  end
+end